fix references rebase merge conflicts
