+++ snake.h
@@ -1,6 +1,11 @@
 #ifndef snake_h
 #define snake_h
 
+#include <unistd.h>
+#include <stdlib.h>
+#include <time.h>
+#include <ctype.h>
+
 #include "position.h"
 
 typedef struct {
@@ -8,7 +13,7 @@ typedef struct {
     Position* tail;
 } Snake;
 
-void initSnake();
+void initSnake(int my);
 void startGame();
 void gameOver();
 void moveHead(Snake* snake, int dir);
